+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - il2cpp remove zeroes from path ([#179](https://github.com/getsentry/sentry-unity/pull/179))
+- Flag simulator based on Application.isEditor ([#184](https://github.com/getsentry/sentry-unity/pull/184))
 
 ### Features